PPhotonic_Radar_AI_Project-master/requirements.txt
# Primary Dependencies
numpy
scipy
matplotlib
streamlit
# PyTorch - Install with: pip install torch --index-url https://download.pytorch.org/whl/cpu
# Or for GPU: pip install torch torchvision torchaudio
opencv-python
psutil
seaborn
scikit-learn
plotly

# Optional (for RTL-SDR hardware support)
# pyrtlsdr
# kafka-python

# Configuration & logging helpers
pyyaml
python-json-logger

# Optional / Legacy
# tensorflow (Not supported on Python 3.14 yet)
